# docs/swagger.yaml
definitions:
  models.Product:
    properties:
      category:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
info:
  contact: {}
paths:
  /products:
    delete:
      consumes:
      - application/json
      description: Delete products matching the specified filter criteria
      parameters:
      - description: Filter Criteria
        in: body
        name: filter
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Products deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid Filter Criteria
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete multiple products
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Retrieve all products from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product with the given details
      parameters:
      - description: Product Data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a product
      tags:
      - products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific product by its unique ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid Product ID
          schema:
            type: string
        "404":
          description: Product Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete a product
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Retrieve a specific product using its unique ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid Product ID
          schema:
            type: string
        "404":
          description: Product Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update the details of an existing product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Product Data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid Product ID or Request Body
          schema:
            type: string
        "404":
          description: Product Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update a product
      tags:
      - products
  /products/{id}/replace:
    put:
      consumes:
      - application/json
      description: Replace the details of an existing product with the provided new
        details
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product Data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: Product replaced successfully
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid Product ID or Request Body
          schema:
            type: string
        "404":
          description: Product Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Replace a product
      tags:
      - products
  /products/bulk:
    post:
      consumes:
      - application/json
      description: Insert multiple products into the database at once
      parameters:
      - description: List of Products
        in: body
        name: products
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Product'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Products uploaded successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid Input
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Bulk insert products
      tags:
      - products
  /products/category/{category}:
    get:
      consumes:
      - application/json
      description: Retrieve a list of products for a given category
      parameters:
      - description: Product Category
        example: electronics
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of products
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Category not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get products by category
      tags:
      - products
swagger: "2.0"